Purchase installments

+++ app/Actions/Schedule/PurchaseInstallment.php
@@ -71,7 +71,7 @@ class PurchaseInstallment
 
         $paymentIntent = $this->stripe->paymentIntents->create(
             [
-                'amount'               => $depositAmount * 100,
+                'amount'               => (int)($depositAmount * 100),
                 'currency'             => config('app.platform_currency'),
                 'payment_method_types' => ['card'],
                 'customer'             => Auth::user()->stripe_customer_id,
@@ -115,7 +115,7 @@ class PurchaseInstallment
                 [
                     'currency'    => config('app.platform_currency'),
                     'product'     => $this->service->stripe_id,
-                    'unit_amount' => $installmentInfo['amountPerPeriod'] * 100,
+                    'unit_amount' => (int)($installmentInfo['amountPerPeriod'] * 100),
                     'recurring'   => ['interval' => 'day', 'interval_count' => $installmentInfo['daysPerPeriod']],
                 ]
             );
